requirements.txt
# FastAPI and Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.44

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

# AWS Integration
boto3==1.34.27

# QR Code Generation
qrcode==7.4.2
pillow

# Environment & Configuration
python-dotenv==1.0.0
pydantic[email]

# CORS
python-multipart==0.0.6